Refactor C code for finding peaks.
This commit tidies up the C code as well as adds a new paremeter for
extracting all peaks in the RI window or selecting whether the
selected peak is the most intense or the closest to the expected RI.
This is controlled by the option filterResult.

Update functions riMatrix and FinPeaks to fit in the changes of the
underlying C functions. No visible changes for the end user.