Correct freeing of C structures
Problem: some allocations are performed by calls to Calloc/Realloc,
while others by the R interface, which means calls to Free needs
to be called when corresponds, otherwise a segfault is raised.

This commit adds a flag to the matrix_t struct to indicate what
function allocated the memory and then free it if necessary,
as well as remove duplicated code in ncdf.c

Finally, fix UNPROTECT calls.