Restructuring of baseline corrrection functions
New baseline correction algorthim based on quantiles around a time
window. The algorithm is applied on every ion trace. The function
accepts a ncdf list or a matrix as input.

Refactor the classic baselineCorrection method such it takes a ncdf
list or a matrix. Note the input matrix is transposed respect to
old TargetSearch versions.

Simplify `baseline` function. Takes either a cdfFile or a ncdf list.
Moreover, a new parameter bsline_method can be passed to chose between
the aforementioned methods.