Refactor parsing of column names of RI files in text format
Many functions require an argument for the column names of the RI files
in text format. This format is mostly obsolete but kept for historical
reasons, and the column names are always the same in any case.

This commit gets rid of the need to set this parameter (it is now NULL
by default), which means that the default names will be used. A new
option called 'TS_RI_columns' can be set globally so we do not worry
about column names. This will be documented in the file `file.Rd`