ri_data_extract: time range can be a matrix
Formerly, the searched time interval (parameter `timeRange`) was limited
to a single ranged for all the searched masses. This commit allows to search
for masses in different time ranges adding more flexibility.

The parameter `timeRange` can be either a matrix or a vector of even length;
it will be coerced to matrix internally.