Make assertion to check for NULL or NA
Rename the function that checks for NULL or NA and make sure that we
check for a scalar equal to NA.

Previously, the function would return TRUE on an input containing NA
such as `c(1, NA, 3)`. This should be handle by another check, so
another check was added.

The function `has_na` was created to check whether the input contains
missing values.