Fix: plotRIdev use near equality for comparison
The function compares the selective or top masses agains the columns
of the intensity matrix. However, their lengths are variable, so the
for `if` condition throwed a warning if the lengths didn't match.

The fix uses the R function `all.equal` to test for near equality
instead, which is more robust for unequal lengths.