Deprecate parameter 'libId' and replace it with 'libID'
The parameter `libId` conflicts with the method libId, and in other
functions we already use a variable `libID` that correspond with the
library identifier. Therefore, we replace it for consistency.

In a subsequent commit, the plotSpectra function will be updated and
this update will take care of the outdated parameter (ie, it will issue
a warning).