Refactor: fix quantMatrix function
The quantMatrix function have been broken for years as it did not
work as intended (options maxobs/maxint were ignored), because
we enforce a quantification mass (before we didn't).

The function will have three methods: 'quantmass' will use the
library's quant mass (this is how the function actually worked),
and then previous methods 'maxint' and 'maxobs' which ignore the
quantification mass.

The output also changes. Before the rownames where the metabolite
names, now they are the library unique identifiers. The metabolite
names are returned as an attribute.

This commit also updates the documentation and adds unit testing.