Refactor: Write.Table function
Due to changes to the function `quantMatrix`, update the function, in
particular the parameter `quantMatrix` to accept the same options of
that function. In addition add the option `selmass` that is passed to
`quantMatrix`

Change the default option of args `prefix` from `NA` to `NULL`, though
the former is still valid. Also add assertions for `prefix`

Update documentation to reflect the changes. Also explain what files
are created and their content.