Bugfix: refactor get_line function
The function can potentially segfault if a file contains CR only
line breaks, which are common in Mac OS systems.

The function checks of all common types of line breaks, namely
LF (linux), CRLF (windows), or CR (Mac OS).

In addition, the function was renamed from `getLine` to `get_line` to
be more consistent with the naming convention. (The name `getline`
conflicts with getline(3) available in GNU libc systems)