Refactor: check for NULL return values
R takes care of allocation error of the user-controlled memory macros
`R_Calloc` and `R_Realloc`, so they do not return `NULL`, so checking
for their return value is redundant. However, this check is necessary
when using the functions directly in C.

When working directly in `C`, the aforementioned macros are implemented
as macros of calloc(3) and realloc(3) respectively, so checking for
`NULL` here is actually needed.