Refactor: Update C functions to manipulate files
The reason for updating the C functions was that static analysis and
fuzzy testing of the C code revealed some vulnerabilities. Another
difficulty is that the code relied on R transient storage memory which I
find hard to debug, in particular if one wants to reuse the code outside
R in a C-based program.

Most of the changes involve new structures and using 'user-controlled
memory' (instead of transient storage) as mentioned. The R interfaces to
C code, i.e., the .C and .Call functions remained unchanged.

All in all, these changes should no be visible to the end user.