Refactor: repurposes `find_peaks` function parameters
Change the order and type of `find_peaks` arguments, in particular,
`Options` is removed and  replaced by `Columns`. The file type and swap
(formerly in `Options`) are detected at runtime using C functions.

The C function `read_file` changes its arguments to `struct column_s`
instead of column names, so it accept indices as well (formerly only
column names where allowed).

The function `getAllPeaks` is moved to its own file (find.R), which
also includes the new R function `.c_find_peaks` which is the interface
to `find_peaks`.

Replace the use of `.Call` in the functions `FindPeaks`, `riMatrix`
and `ri_data_extract` with calls to the aforementioned R to C
interface function.