Refactor: remove function `get.file.format.opt`
The code of this function has been migrated to C, so it is not
needed. Some functionality is exposed by the function `file_type`

Note: the refactoring uncover a bug in `plotPeak` which makes
splitting the fix in another commit impossible. However, this will
be noted in the next commit as the documentation was outdated.