Refactor: Improve `tsLib` spectra handling
Previously, there were no checks on the spectra attribute of objects of
class `tsLib`. This commit makes sures that the spectra is kept
consistent.

This commits also allows for missing spectra on some metabolites, in
which case the spetrum matrix must be set to a vector of zero length.

The internal function `Top.Masses` might get an empty spectrum, so in
this case we return a zero-length numeric vector.