fix: use \providecommand for missing defines in vignettes
There are missing latex definitions due to a bug in BiocStyle
caused by knitr 1.47, which have been fixed in 5ca337b.
This has been reported in #109 there.

However, a recent commit in BiocStyle (a13ba84) add also the same
definitions, so we use \providecommand instead to avoid latex errors
when a new release of BiocStyle hit the repos.

Also, use the same colors of BiocStyle to keep the same style.