Feature: allow incompatible data when combining objects
The objects `tsLib` and `tsSample` allow combination by using the `c`
method. However, it used to fail if the data.frame on the `data` slot
contained incompatible columns.

This commit makes combinations of such objects possible in most cases,
though there are limitations with some data types (like lists or
matrices).

In case the combination is not possible, throw a more meaningful error
message.