fix: FAMEoutliers manual day grouping
There was an unnoticed bug when selecting manual day grouping using the
`startDay` and `endDay` parameters. The reason is it tried to compare
numeric with characters vectors. In addition, it checked for NAs in an
`if` condition with `is.na`, that does not return a scalar.

This commit refactors the check for `NAs`, adds more assertions
to make sure the startDay and endDay parameters match the sample days,
and the comparison logic uses indices instead.

Notice that there could be a broken change since these parameters
default to `NULL` instead of `NA`.

Docs: update man page to reflect those changes and add examples of
the manual grouping.

Finally, refactor on the calculation of the daily mean with `vapply`
(instead of a for loop).