refactor: deprecate method `libId`
Deprecates the ill-named method `libId` and replace it with the
function`makeIndex`. Both functions are equivalent, so functionality is
not loss.

The method is mostly used internally so it should not cause disruption.
The reason of the change is to make space for a new function/method
that deals with the library's identifiers.