refactor: internal function .setLibrary and .set_uid
Split the logic that remove artifacs from the `libData` slot into a
separated function.

Create a new internal function `.set_uid` which takes cares of setting
or getting the unique identifiers for each library element.

Add a test that checks an expected warning, and also check that the
library object is created cleanly.