fix: RIcorrect cannot open file under certain conditions
This occurs when the input files are NC4 and the `writeCDF4path` option is
set to write to a different path. Same if they are TargetSearch files,
but lack the correct extension.

This commit fixes the issue, however, if the input files contain the
extension NC4 and they are not created by TargetSearch, an error will
be thrown in order to avoid accidental overwriting. This is documented
on the man page.